+++ app/src/main/java/merail/life/mejourney/data/HomeItem.kt
@@ -1,7 +1,7 @@
 package merail.life.mejourney.data
 
 data class HomeItem(
-    val year: Int,
+    val year: Long,
     val title: String,
     val description: String,
     val url: String = "",
